로그인 컨트롤러 일부 수정 로그인시 경매게시판 상세 게시물에서 찜하기 여부에 따른 찜취소, 찜추가 가능

+++ HiddenPiece/src/main/java/org/goodomen/hiddenpiece/model/service/MemberService.java
@@ -29,5 +29,5 @@ public interface MemberService {
 
 	void deleteFromWishlist(AuctionBoardLikesVO likesVO);
 
-	long checkWishlist(AuctionBoardLikesVO likesVO);
+	int checkWishlist(AuctionBoardLikesVO likesVO);
 }
